regtypes iq max unsigned 32 bit value overflow fix

+++ regtypes.h
@@ -231,10 +231,10 @@ extern bool isMemory(DataType type);
 extern int32_t iqBase(DataType type);
 
 // Минимальное целочисленное значение заданного типа.
-int32_t iqIntMinimum(DataType type);
+int64_t iqIntMinimum(DataType type);
 
 // Максимальное целочисленное значение заданного типа.
-int32_t iqIntMaximum(DataType type);
+int64_t iqIntMaximum(DataType type);
 
 // Минимальное значение заданного типа.
 // intN - принимается как IQN.0
